auth_service/requirements.txt
fastapi
uvicorn
motor
pydantic
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
pytest
httpx